+++ apps/web/src/app/status/[slug]/page.tsx
@@ -24,7 +24,7 @@ import { last } from "@/lib/utils";
 import { ThemeToggle } from "../../../components/theme-toggle";
 import { useLocalizedTranslation } from "@/hooks/useTranslation";
 
-const PublicStatusPage = ({ incomingSlug = "" }: { incomingSlug?: string }) => {
+const PublicStatusPage = ({ incomingSlug }: { incomingSlug?: string }) => {
   const params = useParams<{ slug: string }>();
   const slug = incomingSlug ?? params.slug;
   const { t } = useLocalizedTranslation();