+++ CHANGELOG.md
@@ -16,6 +16,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - add the server name in message for pushover
 - fix broken FindAll with tags
 
+- Fix custom domain issues
+
 ## [0.0.36] - 2025-07-28
 
 ### Added